requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
openai>=1.40.0
pypdf>=4.3.1
python-docx>=1.1.2
python-multipart>=0.0.9
scikit-learn>=1.5.0
numpy>=1.26.0
celery>=5.4.0
redis>=5.0.7